requirements/base.txt
anyascii==0.3.0
asgiref==3.5.0
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.11
django-filter==2.4.0
django-mjml==0.12.0
django-modelcluster==5.2
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
openpyxl==3.0.9
Pillow==8.4.0
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
urllib3==1.26.8
wagtail==2.15.3
wagtail-birdsong==0.2.0
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
python-dotenv==0.19.2
django-debug-toolbar==3.2.4
wagtail-seo==1.0.0
django-cogwheels==0.3 
wagtailmenus==3.0.2
django-widget-tweaks==1.4.12
defusedxml==0.7.1 
lml==0.1.0 
odfpy==1.4.1 
pyexcel-io==0.6.6 
pyexcel-ods==0.6.0